fix: resolve CI failures — mypy unused-ignore, isort order, brisque float cast
- Remove redundant # type: ignore[import*] comments from calibration.py,
  pdf.py, and cli.py; ignore_missing_imports=true already covers these globally
- Remove unused # type: ignore[type-arg] from cli.py _print_grid_text/_print_pdf_text
- brisque_like.py: cast np.mean(...) to float to fix mypy operator type error
- isort: fix import order in tests/test_positive_orientation.py
All checks now pass: mypy ✓  isort ✓  ruff ✓  black ✓  pytest 161/161 ✓

+++ doc_qual/features/brisque_like.py
@@ -134,7 +134,7 @@ def brisque_like_features(
     # Average pair α deviation from 2.0 — pairwise GGD should also be ≈ 2
     pair_alpha_vals = list(pair_alphas.values())
     pair_score = 100.0 - min(
-        100.0, np.mean([abs(a - 2.0) for a in pair_alpha_vals]) * 35.0
+        100.0, float(np.mean([abs(a - 2.0) for a in pair_alpha_vals])) * 35.0
     )
 
     score = float(0.4 * alpha_score + 0.35 * sigma_score + 0.25 * pair_score)
